pc
Write a program in js to calculate and print the Electricity bill of a given customer. The customer id., full name of the user and unit consumed by the user should be taken from the keyboard and display the total amount to pay to the customer as "Hello XYZ, Customer id: 1234, your bill is: 1234"
The charge are as follow :

Unit	Charge/unit
upto 199      	@1.20


200 and above but less than 400	@1.50


400 and above but less than 600	@1.80



600 and above	@2.00


If bill exceeds Rs. 400 then a surcharge of 15% will be charged and the minimum bill should be of Rs. 100/-